Refactor activate.rs: Implement undo and error handling

+++ src/helper/activate.rs
@@ -1,25 +1,26 @@
 use crate::helper::get_target_data;
 use crate::modules::db;
-use crate::skinners::r;
+use crate::skinners::undo;
+use crate::{error, error::throw};
 
-pub fn action(target: &str) -> Result<(), Box<dyn std::error::Error>> {
+pub fn action(target: &str) -> Result<bool, Box<dyn std::error::Error>> {
 	match get_target_data(target) {
 		Ok(metadata_vec) => {
 			let count = metadata_vec.len();
 			if count == 0 {
-				println!("The command does not exist in the database, or it is activated");
+				throw!(error::TARGET_UNACTIVATABLE);
 			} else {
 				let db_conn = db::connect_to_main_db().expect("Failed to connect to the database");
 				for metadata in metadata_vec {
-					match db::execute_sql(
-						&db_conn,
-						format!("DELETE FROM TARGETS WHERE TARGET = '{}'", metadata.target)
-							.as_str(),
-					) {
-						Ok(_) => {
-							r::undo(metadata);
+					if undo(metadata.clone()) {
+						match db::execute_sql(
+							&db_conn,
+							format!("DELETE FROM TARGETS WHERE TARGET = '{}'", metadata.target)
+								.as_str(),
+						) {
+							Ok(_) => {}
+							Err(err) => println!("Error deleting from database: {:?}", err),
 						}
-						Err(err) => println!("Error deleting from database: {:?}", err),
 					}
 				}
 			}
@@ -27,5 +28,5 @@ pub fn action(target: &str) -> Result<(), Box<dyn std::error::Error>> {
 		Err(err) => println!("Error retrieving target data: {:?}", err),
 	}
 
-	Ok(())
+	Ok(false)
 }
